backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
django-environ>=0.11,<1.0
django-celery-beat>=2.6,<3.0
celery[redis]>=5.4,<6.0
psycopg2-binary>=2.9,<3.0
redis>=5.0,<6.0
python-barcode>=0.15,<1.0
Pillow>=10.0,<11.0
drf-spectacular>=0.27,<1.0
gunicorn>=22.0,<23.0
httpx>=0.27,<1.0
stripe>=11.0,<12.0
dj-stripe>=2.8,<3.0
